U: Controllers divided by entity
