+++ API/WebServer/Controllers/preliminarySessionController.go
@@ -7,27 +7,6 @@ import (
 	"github.com/gofiber/fiber/v2"
 )
 
-func CancelSession(context *fiber.Ctx) error {
-	token := analyzeToken(context)
-
-	if token == nil {
-		return nil
-	}
-
-	var data map[string]string
-	if err := context.BodyParser(&data); err != nil {
-		return err
-	}
-
-	date := data["date"]
-	roomId, _ := strconv.Atoi(data["roomId"])
-	startTime := data["startTime"]
-
-	result := Requests.CancelSession(date, roomId, startTime)
-
-	return giveVoidOperationResponse(context, result)
-}
-
 func GetPreliminarySchedule(context *fiber.Ctx) error {
 	token := analyzeToken(context)
 
@@ -48,27 +27,6 @@ func GetPreliminarySchedule(context *fiber.Ctx) error {
 	return giveJSONResponse(context, dummySchedule, fiber.StatusOK)
 }
 
-func DeletePreliminarySession(context *fiber.Ctx) error {
-	token := analyzeToken(context)
-
-	if token == nil {
-		return nil
-	}
-	var data map[string]string
-	if err := context.BodyParser(&data); err != nil {
-		return err
-	}
-
-	year, _ := strconv.Atoi(data["year"])
-	month, _ := strconv.Atoi(data["month"])
-	weekDay, _ := strconv.Atoi(data["weekDay"])
-	roomId, _ := strconv.Atoi(data["roomId"])
-	startTime := data["startTime"]
-
-	result := Requests.DeletePreliminarySession(year, month, weekDay, roomId, startTime)
-
-	return giveVoidOperationResponse(context, result)
-}
 func InsertPreliminarySession(context *fiber.Ctx) error {
 
 	token := analyzeToken(context)
@@ -120,22 +78,24 @@ func ConfirmPreliminarySchedule(context *fiber.Ctx) error {
 
 }
 
-func DeleteInstructor(context *fiber.Ctx) error {
-
+func DeletePreliminarySession(context *fiber.Ctx) error {
 	token := analyzeToken(context)
 
 	if token == nil {
 		return nil
 	}
-
 	var data map[string]string
 	if err := context.BodyParser(&data); err != nil {
 		return err
 	}
 
-	instructorNumber, _ := strconv.Atoi(data["instructorNumber"])
+	year, _ := strconv.Atoi(data["year"])
+	month, _ := strconv.Atoi(data["month"])
+	weekDay, _ := strconv.Atoi(data["weekDay"])
+	roomId, _ := strconv.Atoi(data["roomId"])
+	startTime := data["startTime"]
 
-	result := Requests.DeleteInstructor(instructorNumber)
+	result := Requests.DeletePreliminarySession(year, month, weekDay, roomId, startTime)
 
 	return giveVoidOperationResponse(context, result)
 }
